requirements.txt
albumentations==1.3.1
asposestorage==1.0.2
data==0.4
faiss_cpu==1.7.4
matplotlib==3.8.2
numpy==1.26.3
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
torch==2.1.2
tqdm==4.66.1
